Refactoring: Move orioledb_undo_size from handler.c to src/transam/undo.c
